173 - persist num files comparison

+++ workers/workers/scripts/verify_archive_paths.py
@@ -1,6 +1,7 @@
 import logging
 import shutil
 from pathlib import Path
+import os
 
 import fire
 import workers.api as api
@@ -12,42 +13,75 @@ logger = logging.getLogger(__name__)
 
 
 def verify():
+    archive_path_verification_dir = Path(config['paths'][dataset['type']]['archive_path_verification'])
+    logger.info(f"Archive path verification dir: {archive_path_verification_dir}")
+    archive_path_verification_dir.mkdir(exist_ok=True, parents=True)
+
     incorrect_path_dataset_ids = []
+    correct_path_dataset_ids = []
+    unprocessed_dataset_ids = []
 
     datasets = api.get_all_datasets(archived=True, bundle=True)
 
     for dataset in datasets:
+        logger.info(f"Verifying dataset {dataset['id']}")
+        
         bundle = dataset['bundle']
+        if bundle is None:
+            logger.info(f"No bundle for dataset {dataset['id']}. Will not process. Skipping...")
+            unprocessed_dataset_ids.append(dataset['id'])
+            continue
+        
         archive_path = dataset['archive_path']
-
-        archive_path_verification_dir = Path(config['paths'][dataset['type']]['archive_path_verification'])
-        archive_path_verification_dir.mkdir(exist_ok=True)
-
+        
         working_dir = archive_path_verification_dir / dataset['name']
         if working_dir.exists():
             shutil.rmtree(working_dir)
+        working_dir.mkdir(exist_ok=True, parents=True)
+
+        logger.info(f"Working dir: {working_dir}")
+        tar_file_download_path = working_dir / bundle['name']
+
+        logger.info(f"Tar file download path: {tar_file_download_path}")
+
 
         wf_utils.download_file_from_sda(sda_file_path=archive_path,
-                                        local_file_path=working_dir)
-        tar_file_path = working_dir / bundle['name']
+                                        local_file_path=tar_file_download_path)
 
-        if tar_file_path.exists():
-            shutil.rmtree(tar_file_path)
+        extraction_path = working_dir / f"{dataset['name']}_extracted" / dataset['name']
+        logger.info(f"Extraction path: {extraction_path}")
 
-        extraction_path = working_dir / 'dummy_path'
-        wf_utils.extract_tarfile(tar_path=tar_file_path, target_dir=extraction_path)
+        wf_utils.extract_tarfile(tar_path=tar_file_download_path, target_dir=extraction_path, override_arcname=True)
 
-        logger.info(f"Dataset {dataset['id']} extraction path: {extraction_path}")
+        dataset_origin_path = config['registration'][dataset['type']]['source_dir']
+        logger.info(f"Dataset origin path: {dataset_origin_path}")
 
-        incorrect_nested_path = extraction_path / config['registration'][dataset['type']]['source_dir'][                                                               1:] / dataset["name"]
+        incorrect_nested_path = extraction_path / dataset_origin_path[1:] / dataset["name"]
+        logger.info(f"Looking for incorrect nested path: {incorrect_nested_path}")
+        
         has_incorrect_path = incorrect_nested_path.exists()
+        logger.info(f"Has incorrect path: {has_incorrect_path}")
 
+        if not has_incorrect_path:
+            extracted_dirs = [x[0] for x in os.walk(extraction_path)]
+            logger.info(f"Extracted dataset dirs: {extracted_dirs}")
+        
         if has_incorrect_path:
-            logger.info(f"Dataset {dataset['name']} has incorrect nested path: {incorrect_nested_path}")
+            logger.info(f"Dataset {dataset['name']} has incorrect nested path {incorrect_nested_path}")
             incorrect_path_dataset_ids.append(dataset['id'])
+        else:
+            logger.info(f"Dataset {dataset['name']} has no incorrect nested path {incorrect_nested_path}")
+            correct_path_dataset_ids.append(dataset['id'])
 
-        logger.info(f"Incorrect path dataset IDs: {incorrect_path_dataset_ids}")
+        shutil.rmtree(working_dir)
+        logger.info(f"Removed working dir: {working_dir}")
 
+    logger.info(f"Incorrect path dataset IDs: {incorrect_path_dataset_ids}")
+    logger.info(f"Correct path dataset IDs: {correct_path_dataset_ids}")
+    logger.info(f"Unprocessed dataset IDs: {unprocessed_dataset_ids}")
+
+    shutil.rmtree(archive_path_verification_dir)
+    logger.info(f"Removed archive path verification dir: {archive_path_verification_dir}")
 
 
 if __name__ == "__main__":
